resources/bin/konatagger/pyproject.toml
[project]
name = "konatagger"
version = "0.1.0"
description = "Tagging images API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "dynaconf>=3.2.7",
  "einops>=0.7.0",
  "fastapi>=0.115.7",
  "pillow>=9.4.0",
  "pillow-avif-plugin>=1.5.0",
  "python-multipart>=0.0.20",
  "safetensors>=0.4.1",
  "torch>=2.0.1",
  "torchvision>=0.15.2",
  "transformers>=4.36.2",
  "uvicorn>=0.34.0",
]

[project.optional-dependencies]
cpu = ["torch>=2.0.1", "torchvision>=0.15.2"]
cu121 = ["torch>=2.0.1", "torchvision>=0.15.2"]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
conflicts = [[{ extra = "cpu" }, { extra = "cu121" }]]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu121", extra = "cu121" },
  { index = "pytorch-cpu", extra = "cpu" },
]
torchvision = [
  { index = "pytorch-cu121", extra = "cu121" },
  { index = "pytorch-cpu", extra = "cpu" },
]
